% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Trading Algorithm.}
\usage{
format_data(pt_name, pt_amount, pt_price)
}
\arguments{
\item{producers}{producers dataframe.}

\item{users}{users dataframe.}

\item{deal_ratio_proc}{deal ratio in producers.}
}
\value{
\code{data} which is extracted from extracted.
}
\description{
Trading Algorithm.
}

