% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_trading.R
\name{main_trading}
\alias{main_trading}
\title{Trading Algorithm.}
\usage{
main_trading(producers, users, deal_ratio_proc = 0.8)
}
\arguments{
\item{producers}{producers dataframe.}

\item{users}{users dataframe.}

\item{deal_ratio_proc}{deal ratio in producers.}
}
\value{
\code{data} which is extracted from extracted.
}
\description{
Trading Algorithm.
}

